Centralizando metodos en App
