+++ app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".DatosApp"
+        android:name=".App"
         android:allowBackup="true"
         android:icon="@drawable/iconsquare"
         android:label="@string/app_name"
