+++ app/src/main/java/com/example/inkler/FichaEstudio.java
@@ -38,7 +38,6 @@ public class FichaEstudio extends AppCompatActivity {
     private RecyclerView.LayoutManager layoutManagerWeb;
     private AdaptadorTatuadorWeb adaptadorWeb;
     private MapView mapView;
-    private MetodosComunes metodosComunes;
     private DBlocal db;
     RecyclerView recyclerView;
     private RecyclerView.LayoutManager layoutManager;
@@ -50,7 +49,6 @@ public class FichaEstudio extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         db = new DBlocal(getApplicationContext());
-        metodosComunes = new MetodosComunes();
         final int idEstudio = getIntent().getIntExtra("idEstudio",0);
         estudio = db.recogerEstudio(idEstudio);
         final int INITIAL_ZOOM = 14;
@@ -175,7 +173,7 @@ public class FichaEstudio extends AppCompatActivity {
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
         getMenuInflater().inflate(R.menu.menu_actions, menu);
-        if (DatosApp.isAdmin()) {
+        if (App.isAdmin()) {
             menu.setGroupVisible(R.id.añadir, true);
             menu.setGroupVisible(R.id.modificar, true);
             menu.setGroupVisible(R.id.logout, true);
@@ -210,14 +208,14 @@ public class FichaEstudio extends AppCompatActivity {
                 public void onClick(DialogInterface dialog, int which) {
                     String password = input.getText().toString();
                     if (getString(R.string.contraseña).equals(password)){
-                        DatosApp.setAdmin(true);
+                        App.setAdmin(true);
                         invalidateOptionsMenu();
                     }
                 }
             });
             alertDialog.show();
         } else if (id == R.id.noadmin) {
-            DatosApp.setAdmin(false);
+            App.setAdmin(false);
             invalidateOptionsMenu();
         }
         else if (id == R.id.añadir_tatuador) {
